Title: 컴퓨터와 함께하는 숫자 게임 2, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 240307/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.java
@@ -13,11 +13,11 @@ public class Main {
         int a = Integer.parseInt(st.nextToken());
         int b = Integer.parseInt(st.nextToken());
 
-        int max = Integer.MIN_VALUE;
-        int min = Integer.MAX_VALUE;
+        long max = Long.MIN_VALUE;
+        long min = Long.MAX_VALUE;
 
         for(int i = a; i <= b; i++){
-            int res = binarySearch(i);
+            long res = binarySearch(i);
             if(max < res){
                 max = res;
             }
@@ -31,10 +31,10 @@ public class Main {
 
     }
 
-    public static int binarySearch(int target){
+    public static long binarySearch(int target){
         long left = 1;
         long right = m;
-        int cnt = 1;
+        long cnt = 1;
 
         while(left <= right){
             long mid = (left+right)/2;
